Revert "feat(common): LFG-46 Add mysql and heroku support"
