+++ .env-sample
@@ -22,11 +22,3 @@ DB_TYPE=firebase
 FIRE_API_KEY={firebase key}
 FIRE_DOMAIN={firebase domain}
 FIRE_PROJECT_ID={firebase project id}
-
-# If using mysql, enter your config here
-
-MYSQL_HOST={mysql host}
-MYSQL_DATABASE={mysql domain}
-MYSQL_USERNAME={mysql username}
-MYSQL_PASSWORD={mysql password}
-MYSQL_PORT={mysql port}
