--- app.json
@@ -1,34 +1,0 @@
-{
-  "name": "BigCommerce Sample App",
-  "description": "A sample NodeJS app installable on BigCommerce stores",
-  "repository": "https://github.com/bigcommerce/sample-app-nodejs",
-  "keywords": ["Node", "Heroku"],
-  "addons": ["cleardb:ignite"],
-  "env": {
-    "CLIENT_ID": {
-      "description": "The client id of your app, provided by the BigCommerce developer portal."
-    },
-    "CLIENT_SECRET":{
-      "description": "The client secret of your app, provided by the BigCommerce developer portal."
-    },
-    "AUTH_CALLBACK":{
-      "description": "The fully qualified authentication endpoint provided by this app. Replace the app name with the app name chosen above.",
-      "value": "https://<APP NAME>.herokuapp.com/api/auth"
-    },
-    "DB_TYPE":{
-      "description": "Which type of database we are using. Options are 'mysql' or 'firebase'. Can safely leave unchanged.",
-      "value": "mysql"
-    },
-    "COOKIE_NAME":{
-      "description": "The cookie name to use for storing session data in the browser. Can safely leave unchanged.",
-      "value": "_bcsampleapp"
-    },
-    "JWT_KEY":{
-      "description": "Key the app will use for signing JWT. Can safely leave unchanged.",
-      "generator": "secret"
-    }
-  },
-  "scripts": {
-    "postdeploy": "npm run db:setup"
-  }
-}
