--- lib/db.ts
@@ -1,19 +1,0 @@
-import { Db } from '../types'
-
-const { DB_TYPE } = process.env;
-
-let db: Db;
-
-switch (DB_TYPE) {
-    case 'firebase':
-        db = require('./dbs/firebase');
-        break;
-    case 'mysql':
-        db = require('./dbs/mysql');
-        break;
-    default:
-        db = require('./dbs/firebase');
-        break;
-}
-
-export default db;
