+++ package.json
@@ -6,10 +6,9 @@
   "scripts": {
     "dev": "next",
     "build": "next build",
-    "start": "next start -p $PORT",
+    "start": "next start",
     "test": "jest",
-    "lint": "eslint . --ext .ts,.tsx,.js",
-    "db:setup": "node scripts/db.js"
+    "lint": "eslint . --ext .ts,.tsx,.js"
   },
   "keywords": [],
   "author": "",
@@ -19,7 +18,6 @@
     "cookie": "^0.4.1",
     "firebase": "^8.2.6",
     "jsonwebtoken": "^8.5.1",
-    "mysql": "^2.18.1",
     "next": "^10.0.5",
     "node-bigcommerce": "^4.1.0",
     "react": "^17.0.1",
