+++ types/index.ts
@@ -1,3 +1,3 @@
 export * from './auth';
 export * from './data';
-export * from './db';
+export * from './firebase';
